% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeNeighborTiles.R
\name{checkCoordinates}
\alias{checkCoordinates}
\title{Check coordinates}
\usage{
checkCoordinates(spatialDF)
}
\arguments{
\item{spatialDF}{a spatiald data frame}
}
\description{
Function checks if there is a spatialdataframe
}
\details{
Check coords
}
