% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkIfMultiTile.R
\name{checkIfMultiTile}
\alias{checkIfMultiTile}
\title{Check multitile}
\usage{
checkIfMultiTile(path, coordX, coordY)
}
\arguments{
\item{path}{location to the files}

\item{coordX}{x-coord name of the file, rounded 6 numbered RD-coord}

\item{coordY}{y-coord name of the file, rounded 6 numbered RD-coord}
}
\description{
Check neighboring tiles
}
