% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extent_from_new_pro.R
\name{extent_from_new_pro}
\alias{extent_from_new_pro}
\title{New Extent}
\usage{
extent_from_new_pro(ext, pro.old = WGS84, pro.new = pro)
}
\arguments{
\item{ext}{the extent in the current projection}

\item{pro.old}{the current projection}

\item{pro.new}{the new projection}
}
\description{
Extent from new projection
}
