% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_file_name_from_extent.R
\name{grid_file_name_from_extent}
\alias{grid_file_name_from_extent}
\title{Grid file names}
\usage{
grid_file_name_from_extent(ext = Haag.ext)
}
\arguments{
\item{ext}{extent in RD-coordinates}
}
\description{
get the grid file names for an extention
}
\details{
Grid file names from extent
}
