% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines_from_las.R
\name{lines_from_las}
\alias{lines_from_las}
\title{Calculate pahts from a raster}
\usage{
lines_from_las(LAS = testTile, pro = CRS("+init=epsg:28992"), xres = 5,
  yres = 5, field = "Z", point = testPoint, R = 200,
  Runcertainty = 10, theta = c(5, 10, 30))
}
\arguments{
\item{LAS}{location of the las file to read}

\item{pro}{projection, now only 28992 (RDcoordinates) is excepted}

\item{xres}{x-resolution to grid the las points}

\item{yres}{y-resolution to grid the las points}

\item{field}{field to grid}

\item{point}{location for which the lines are calculated}

\item{R}{radius for which the line is calculated}

\item{Runcertainty}{can be used if the location coordinates are uncertain or if you want a surrounding area}

\item{theta}{angles for which the path is calculated}
}
\description{
\code{calculate_path} calculates the path for given distance, angles and coordinates from a raster file. A test file in
the folder test_data with test parameters is included.
}
\seealso{
This function calls \code{\link{calculate_path}}
}
