% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadtileGMSFromLas.R
\name{loadTileGMSFromLas}
\alias{loadTileGMSFromLas}
\title{Load LAS tile}
\usage{
loadTileGMSFromLas(path, coordX, coordY)
}
\arguments{
\item{path}{path to the files}

\item{coordX}{x-coordinate in RD coordinates}

\item{coordY}{y-coordinate in RD coordinates}
}
\description{
Load a tile
}
