% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadTileWholeNL.R
\name{loadTileWholeNL}
\alias{loadTileWholeNL}
\title{Load Tiles}
\usage{
loadTileWholeNL(filepath)
}
\arguments{
\item{filepath}{path to the LAS files}
}
\description{
Load tiles for the whole Netherlands
}
\details{
Load tile whole Netherlands
}
