% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRaster.R
\name{makeRaster}
\alias{makeRaster}
\title{Make a raster}
\usage{
makeRaster(spatialDF, xres, yres, pro)
}
\arguments{
\item{spatialDF}{a spatialpointsdataframe object}

\item{xres}{resolution in the x direction}

\item{yres}{resolution in the y-direction}

\item{pro}{the projection of the raster}
}
\description{
Create a raster
}
