% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSpatialDF.R
\name{makeSpatialDF}
\alias{makeSpatialDF}
\title{Create a SpatialPointDataFrmae}
\usage{
makeSpatialDF(df, projection)
}
\arguments{
\item{df}{a dataframe with X and Y as coordinates}

\item{projection}{the projection to add to the dataframe}
}
\description{
Make a spatialpointsdataframe
}
