% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeNeighborTilesGSMTest.R
\name{mergeNeighborTilesGSMTest}
\alias{mergeNeighborTilesGSMTest}
\title{Merge Tiles test function}
\usage{
mergeNeighborTilesGSMTest(path, tileNumberXCoord, tileNumberYCoord,
  extensionMainTile, maxView, projection)
}
\arguments{
\item{path}{path to the tiles}

\item{maxView}{?}

\item{projection}{a projection, guess RDcoords?}

\item{tileNumberXcoord}{X-coordinates of the center tile}

\item{tileNumberYcoord}{Y-coordinates of the center tile}

\item{extentionMainTile}{no clue, Andrea?}
}
\description{
Merge tiles
}
