% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_over_grid.R
\name{point_info_from_grid}
\alias{point_info_from_grid}
\title{Point info from grid}
\usage{
point_info_from_grid(point = df[1, ], n = 2,
  save_dir = "data/test_point.txt")
}
\arguments{
\item{point}{a point with tileNumerXCoord, tilenumberYcoord}

\item{n}{layer number of the grid}

\item{save_dir}{save file name with extention .txt}
}
\description{
Point information from the grid files
}
\author{
Marieke Dirksen
}
