% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVF.R
\name{SVF}
\alias{SVF}
\title{Sky view factor}
\usage{
SVF(pointX, pointY, maxView, proj)
}
\arguments{
\item{pointX}{x-coordinates}

\item{pointY}{y-coordinates}

\item{maxView}{viewing distance of the sky view factor}

\item{proj}{projection}
}
\description{
calculates the SVF
}
\details{
\code{SVF} Calculates the sky view factor from las point data and save raster files.
}
\seealso{
This function calls upon the following functions \code{\link{loadTile}}, \code{\link{makeSpatialDF}},
\code{\link{mergeNeighborTiles}}, \code{\link{makeRaster}},and from the horizon package \code{\link{svf}}.
}
