\name{svfAHN2point-package}
\alias{svfAHN2point-package}
\alias{svfAHN2point}
\docType{package}
\title{
\packageTitle{svfAHN2point}
}
\description{
\packageDescription{svfAHN2point}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{svfAHN2point}
\packageIndices{svfAHN2point}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{svfAHN2point}

Maintainer: \packageMaintainer{svfAHN2point}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
