% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVFGMStest.R
\name{SVFGMSTest}
\alias{SVFGMSTest}
\title{Test function for a point dataset of the GMS}
\usage{
SVFGMSTest(pointX, pointY, maxView, proj)
}
\arguments{
\item{pointX}{x-coordinates of the point in RD coordinates}

\item{pointY}{y-coordinates of the point in RD coordinates}

\item{proj}{projection}
}
\description{
Test function
}
