% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVFWholeNL.R
\name{SVFWholeNL}
\alias{SVFWholeNL}
\title{SVF Netherlands}
\usage{
SVFWholeNL(filepath, maxView)
}
\arguments{
\item{filepath}{path to the LAS files}

\item{maxView}{maximum viewing distance for the sky view factor calculations}
}
\description{
Calculate the SVF for the whole Netherlands
}
