% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_path.R
\name{calculate_path}
\alias{calculate_path}
\title{Calculate Path
\code{calculate_path} calculates the path for given distance, angles and coordinates from a raster file.}
\usage{
calculate_path(theta, x = x, y = y, distance, rbuffer)
}
\arguments{
\item{theta}{angle or angles in degrees}

\item{x}{x-coordinates in meters (RDcoordinates)}

\item{y}{y-coordinates in meters (RDcoordinates)}

\item{distance}{distance for which the path should be calculated}

\item{rbuffer}{raster which is masked in \code{\link{lines_from_las}}}
}
\description{
calculates the path for different angles
}
\details{
Calculate paths from a raster
}
\seealso{
This function is called in \code{\link{lines_from_las}}
}
