% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadtile.R
\name{loadTile}
\alias{loadTile}
\title{Load LAS tile}
\usage{
loadTile(path, coordX, coordY)
}
\arguments{
\item{path}{is the path to the files}

\item{coordX}{x-coordinate in RD coordinates}

\item{coordY}{y-coordinate in RD coordinates}
}
\description{
loads the first tile where the point is located
}
\details{
Load a tile
}
