% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeNeighborTiles.R
\name{mergeNeighborTiles}
\alias{mergeNeighborTiles}
\title{Merge Tiles function}
\usage{
mergeNeighborTiles(path, tileNumberXCoord, tileNumberYCoord, extensionMainTile,
  maxView, projection)
}
\arguments{
\item{path}{path to the tiles}

\item{tileNumberXCoord}{X-coordinates of the center tile}

\item{tileNumberYCoord}{Y-coordinates of the center tile}

\item{extensionMainTile}{no clue, Andrea?}

\item{maxView}{svf view}

\item{projection}{a projection, guess RDcoords?}
}
\description{
Merging tiles surrounding the point
}
\details{
Merge tiles
}
