% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeNeighborTilesGSMTest.R
\name{mergeNeighborTilesGSMTest}
\alias{mergeNeighborTilesGSMTest}
\title{Merge Tiles test function}
\usage{
mergeNeighborTilesGSMTest(path, tileNumberXCoord, tileNumberYCoord,
  extensionMainTile, maxView, projection)
}
\arguments{
\item{path}{path to the tiles}

\item{tileNumberXCoord}{X-coordinates of the center tile}

\item{tileNumberYCoord}{Y-coordinates of the center tile}

\item{extensionMainTile}{no clue, Andrea?}

\item{maxView}{in svf calculations}

\item{projection}{a projection, guess RDcoords?}
}
\description{
merging tiles surrounding your object
}
\details{
Merge tiles
}
