% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_over_grid.R
\name{point_info_from_grid}
\alias{point_info_from_grid}
\title{Point info from grid}
\usage{
point_info_from_grid(point = df[1, ], n = 2)
}
\arguments{
\item{point}{a point with x-coords and y-coords in RD}

\item{n}{layer number of the grid}
}
\description{
returns the grid file name the point is in
}
\details{
Point information from the grid files
}
\author{
Marieke Dirksen
}
